+++ main-server/src/main/java/ru/practicum/event/dto/UpdateEventUserRequest.java
@@ -22,7 +22,7 @@ public class UpdateEventUserRequest {
 
     @Size(min = 20, max = 2000)
     String annotation;
-    Integer category;
+    Long category;
     @Size(min = 20, max = 7000)
     String description;
     @FutureOrPresent
